requirements.txt
playwright>=1.40.0
APScheduler>=3.11.1
python-dotenv>=1.0.0
requests>=2.30.0
Pillow>=10.0.0
